+++ lib/src/navigator/navigator_page_lifecycle_mixin.dart
@@ -77,10 +77,8 @@ mixin NavigatorPageLifecycleMixin<T extends StatefulWidget> on State<T> {
     );
 
     _anchors = NavigatorPage.routeSettingsListOf(context);
-    // 如果第一个刚好是在 PageView 上，需要移除掉
-    if (_anchors.first.name == _current.name) {
-      _anchors.removeAt(0);
-    }
+    // 链路上重复的 settings 要去掉
+    _anchors.removeWhere((final it) => it.name == _current.name);
 
     for (final callback in _anchorsObserverCallbacks) {
       callback();
@@ -144,12 +142,10 @@ class _AnchorLifecycleObserver with NavigatorPageObserver {
     final void Function(RouteSettings) callback,
     final RouteSettings routeSettings,
   ) {
-    if (_anchor.name != routeSettings.name ||
-        _delegate._current.isSelected == false) {
+    if (_anchor.name != routeSettings.name || _delegate._current.isSelected == false) {
       return;
     }
-    final idx = _delegate._anchors
-        .indexWhere((final it) => it.name == routeSettings.name);
+    final idx = _delegate._anchors.indexWhere((final it) => it.name == routeSettings.name);
     final ins = _delegate._anchors.sublist(0, idx);
     if (ins.every((final it) => it.isSelected == true)) {
       callback(_delegate._current);